fix: Fix compile error cause ScopeContainer reference
